updated views and templates to show Rules&Routes
Dashboard was updated to show the list of Rules (instead of Routes used
previously) that can contain one or more Routes.
New detail page rule_details.html was created as a clone of
route_details.html.
If a Rule contains only single Route, the statistics (and graphs) are
shown directly, otherwise, a separate pages are used.